+++ app/containers/Transactions/actions.js
@@ -39,10 +39,11 @@ export function loadTransactions() {
  *
  * @return {object} An action object with a type of LOAD_TRANSACTIONS_SUCCESS passing the transactions
  */
-export function transactionsLoaded(transactions) {
+export function transactionsLoaded(transactions, pages) {
   return {
     type: LOAD_TRANSACTIONS_SUCCESS,
     transactions,
+    pages,
   };
 }
