+++ app/utils/injectReducer.js
@@ -21,7 +21,6 @@ export default ({ key, reducer }) => (WrappedComponent) => {
 
     componentWillMount() {
       const { injectReducer } = this.injectors;
-
       injectReducer(key, reducer);
     }
